+++ RELEASES.md
@@ -60,6 +60,7 @@ Upcoming
 - Improve special class loaders error handling
 - Instantiate all query senders with Spring (make Spring aware of all database actions) (make defaultDataSource of example webapp work for all database actions)
 - Add hideMethod attribute for masking strings in ErrorStore and MessageLog 
+- Add hostname property to AppConstants properties